[core]  CORE Change  - Device now use double value

+++ opendevice-core/src/main/java/br/com/criativasoft/opendevice/core/DefaultCommandProcessor.java
@@ -76,7 +76,7 @@ public class DefaultCommandProcessor {
             DeviceCommand deviceCommand = (DeviceCommand) command;
 
             int deviceID = deviceCommand.getDeviceID();
-            long value = deviceCommand.getValue();
+            double value = deviceCommand.getValue();
 
             Device device = manager.findDeviceByUID(deviceID);
